+++ configs/react.js
@@ -21,6 +21,7 @@ module.exports = {
     // https://www.npmjs.com/package/eslint-plugin-react-hooks
     'react-hooks/rules-of-hooks': 'error',
     'react-hooks/exhaustive-deps': 'warn',
+    'react/button-has-type': 'warn',
     'react/jsx-curly-brace-presence': [
       'warn',
       { props: 'never', children: 'never', propElementValues: 'always' },