+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - ... <!-- Add new lines here. -->
 - feat: Relax `react/react-in-jsx-scope` with `'off'`
+- feat: Enable `react/button-has-type` with `'warn'`
 
 ## 8.4.0