+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - feat: Relax `react/react-in-jsx-scope` with `'off'`
 - feat: Enable `react/button-has-type` with `'warn'`
 - feat: Enable `@typescript-eslint/naming-convention` with `'warn'`
+- feat: Update dependencies to latest versions
 
 ## 8.4.0