+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. -->
+- feat: Relax `react/react-in-jsx-scope` with `'off'`
+- feat: Enable `react/button-has-type` with `'warn'`
+- feat: Enable `@typescript-eslint/naming-convention` with `'warn'`
+- feat: Update dependencies to latest versions
 
 ## 8.4.0